requirements.txt
azure-core==1.34.0
azure-identity==1.22.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
waitress==3.0.2
Werkzeug==3.1.3